minitor/requirements.txt
pandas
scikit-learn
lightgbm
joblib
google-cloud-storage
google-cloud-bigquery